GafaelfawrUserMenu takes items as children
Note that the logout button is always present.
